+++ changelog.md
@@ -1,5 +1,9 @@
 ﻿# GeoTagNinja Changelog
 
+v0.5.82xx [202208xx]
+- Added check for exifTool version change tracking
+- Added check for GeoTagNinja version change tracking
+
 v0.5.8251 [20220804]
 - Added "Remove GeoData" to main form as well as Edit form
 - Added this changelog.md file